Integrated TGL in app

+++ tgl/memorys.py
@@ -115,7 +115,7 @@ class MailBox():
                     perm = torch.arange(inv.size(0), dtype=inv.dtype, device=inv.device)
                     perm = inv.new_empty(uni.size(0)).scatter_(0, inv, perm)
                     nid = nid[perm]
-                    mail = mail[perm]
+                    mail = mail[perm].float()
                     mail_ts = mail_ts[perm]
                     self.mailbox[nid.long(), self.next_mail_pos[nid.long()]] = mail
                     self.mailbox_ts[nid.long(), self.next_mail_pos[nid.long()]] = mail_ts
